SYNC-1549 Another unnecessary throw/catch

+++ modules/util/sync-engine/src/main/java/com/liferay/sync/engine/util/FileUtil.java
@@ -117,6 +117,10 @@ public class FileUtil {
 	public static void deleteFile(final Path filePath, boolean retry)
 		throws IOException {
 
+		if ((filePath == null) || Files.notExists(filePath)) {
+			return;
+		}
+
 		try {
 			Files.deleteIfExists(filePath);
 		}
